requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
requests==2.31.0
python-dotenv==1.0.1
langchain==0.1.9
langchain-openai==0.0.5
browser-use==0.1.0
playwright==1.42.0
Pillow==10.2.0
pydantic==2.6.1
nest-asyncio==1.6.0 